Made New Branch, Renamed Basic.h, Updated kernel.c
