+++ kernel/commencom/Common.h
@@ -1,4 +1,5 @@
-#include "Basic.h"
+#include "Stdio.h"
+#include "TODO.h"
 void execute_command(char *input) {
     if (compare_string(input, "EXIT") == 0) {
         // clears the screen to make it seem like it just shut down
