+++ kernel/kernel.c
@@ -1,11 +1,4 @@
-#include "../cpu/idt.h"
-#include "../cpu/isr.h"
-#include "../cpu/timer.h"
-#include "../drivers/display.h"
-#include "../drivers/keyboard.h"
-#include "util.h"
-#include "mem.h"
-#include "commencom/TODO.h"
+#include "commencom/Stdio.h"
 #include "commencom/Common.h"
 /*
                 TODO
